+++ screens/setting/index.js
@@ -2,16 +2,18 @@
  * @Author: czy0729
  * @Date: 2019-05-24 01:34:26
  * @Last Modified by: czy0729
- * @Last Modified time: 2019-05-24 23:34:32
+ * @Last Modified time: 2019-05-26 20:31:43
  */
 import React from 'react'
-import { ScrollView, AsyncStorage } from 'react-native'
+import { ScrollView, AsyncStorage, Alert } from 'react-native'
 import { CacheManager } from 'react-native-expo-image-cache'
-import { Modal, Switch } from '@ant-design/react-native'
-import { Text } from '@components'
+import { Switch } from '@ant-design/react-native'
+import { Popover, Menu, Text } from '@components'
 import { systemStore, userStore } from '@stores'
+import { open } from '@utils'
 import { withHeader, observer } from '@utils/decorators'
 import { info } from '@utils/ui'
+import { IOS, GITHUB_URL } from '@constants'
 import { MODEL_SETTING_QUALITY } from '@constants/model'
 import _ from '@styles'
 import Item from './item'
@@ -24,61 +26,45 @@ class Setting extends React.Component {
     title: '设置'
   }
 
-  setQuality = () => {
-    const data = MODEL_SETTING_QUALITY.data.map(({ label }) => ({
-      text: label,
-      onPress: () => systemStore.setQuality(label)
-    }))
-    Modal.operation(data)
+  setQuality = label => {
+    if (label) {
+      systemStore.setQuality(label)
+    }
   }
 
   clearStorage = () => {
-    Modal.alert(
-      '确定清除缓存?',
-      <Text style={_.mt.xs} type='sub' size={14}>
-        包括图片缓存和页面接口的数据缓存
-      </Text>,
-      [
-        {
-          text: '取消',
-          style: {
-            color: _.colorSub
-          }
-        },
-        {
-          text: '确定',
-          style: {
-            color: _.colorDanger
-          },
-          onPress: async () => {
-            await AsyncStorage.clear()
-            await CacheManager.clearCache()
-            systemStore.setStorage('setting')
-            userStore.setStorage('accessToken')
-            userStore.setStorage('userInfo')
-            info('已清除')
-          }
-        }
-      ]
-    )
-  }
-
-  logout = () => {
-    Modal.alert('确定退出登录?', <Text style={{ marginTop: -_.xs }} />, [
+    Alert.alert('提示', '确定清除缓存? 包括图片缓存和页面接口的数据缓存', [
       {
         text: '取消',
-        style: {
-          color: _.colorSub
-        }
+        style: 'cancel'
       },
       {
         text: '确定',
-        style: {
-          color: _.colorDanger
-        },
+        onPress: async () => {
+          await AsyncStorage.clear()
+          await CacheManager.clearCache()
+          systemStore.setStorage('setting')
+          userStore.setStorage('accessToken')
+          userStore.setStorage('userInfo')
+          userStore.setStorage('userCookie')
+          info('已清除')
+        }
+      }
+    ])
+  }
+
+  logout = () => {
+    const { navigation } = this.props
+    Alert.alert('提示', '确定退出登录?', [
+      {
+        text: '取消',
+        style: 'cancel'
+      },
+      {
+        text: '确定',
         onPress: async () => {
           await userStore.logout()
-          info('已登出')
+          navigation.popToTop()
         }
       }
     ])
@@ -86,15 +72,30 @@ class Setting extends React.Component {
 
   render() {
     const { quality, cnFirst, autoFetch } = systemStore.setting
+
+    const data = MODEL_SETTING_QUALITY.data.map(({ label }) => label)
+    const popoverProps = IOS
+      ? {
+          overlay: <Menu data={data} onSelect={this.setQuality} />
+        }
+      : {
+          data,
+          onSelect: this.setQuality
+        }
     return (
       <ScrollView style={_.container.screen}>
         <Item
           style={_.mt.md}
           hd='图片质量'
-          ft={MODEL_SETTING_QUALITY.getLabel(quality)}
+          ft={
+            <Popover placement='bottom' {...popoverProps}>
+              <Text size={16} type='sub'>
+                {MODEL_SETTING_QUALITY.getLabel(quality)}
+              </Text>
+            </Popover>
+          }
           arrow
           highlight
-          onPress={this.setQuality}
         />
         <Item
           border
@@ -104,15 +105,25 @@ class Setting extends React.Component {
         />
         <Item
           border
-          hd='切换页面时自动请求'
+          hd='优化请求量(部分页面需手动刷新)'
           ft={
             <Switch
-              checked={autoFetch}
+              checked={!autoFetch}
               onChange={systemStore.switchAutoFetch}
             />
           }
           withoutFeedback
         />
+
+        <Item style={_.mt.md} hd='检测更新' ft='开发中' />
+        <Item
+          border
+          hd='项目主页'
+          arrow
+          highlight
+          onPress={() => open(GITHUB_URL)}
+        />
+
         <Item
           style={_.mt.md}
           hd='清除缓存'
@@ -120,6 +131,7 @@ class Setting extends React.Component {
           highlight
           onPress={this.clearStorage}
         />
+
         <Item
           style={_.mt.md}
           hd='退出登录'
